tests/submission/Infraware/Forms/contents/Forms/keygen_willvalidate.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>keygen_willValidate</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
	 <keygen name='key' id='keygen_id'></keygen>
	</form>
	<script>

		var keygen = document.getElementById("keygen_id");

		if (typeof(keygen.willValidate) == "boolean") {
			test(function() {
				assert_equals(keygen.willValidate, false, "willValidate attribute is not correct.");
			});
		} else {
			test(function() {
				assert_unreached("willValidate attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
